requirements.txt
google_adk
python-dotenv
fastapi
uvicorn
aiosqlite
google-genai
pydantic